requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
google-genai==0.2.2
